ENG-728, ENG-732, ENG-744, ENG-693, ENG-389
git-svn-id: https://svn.terracotta.org/repo/tc/dso/trunk@24179 7fc7bbf3-cf45-46d4-be06-341739edd864

+++ dso-l2/src/main/java/com/tc/objectserver/handler/ApplyTransactionChangeHandler.java
@@ -108,7 +108,9 @@ public class ApplyTransactionChangeHandler extends AbstractEventHandler {
 
     if (atc.needsApply()) {
       transactionManager.apply(txn, atc.getObjects(), applyInfo, this.instanceMonitor);
-      garbageCollectionManager.deleteObjects(applyInfo.getObjectIDsToDelete(), atc.allCheckedOutObjects());
+      if ( applyInfo.hasObjectsToDelete() ) {
+        garbageCollectionManager.deleteObjects(applyInfo.getObjectIDsToDelete(), atc.allCheckedOutObjects());
+      }
       txnObjectMgr.applyTransactionComplete(applyInfo);
     } else {
       transactionManager.loadApplyChangeResults(txn, applyInfo);
